chore: simplify utils
